--- Gradle/Core/src/main/java/HelloWorld.java
@@ -1,48 +1,0 @@
-/*
- *  Copyright 2023, TeamDev. All rights reserved.
- *
- *  Redistribution and use in source and/or binary forms, with or without
- *  modification, must retain the above copyright notice and the following
- *  disclaimer.
- *
- *  THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS
- *  "AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT
- *  LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR
- *  A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT
- *  OWNER OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL,
- *  SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT
- *  LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
- *  DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY
- *  THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
- *  (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
- *  OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
- */
-
-import static com.teamdev.jxbrowser.engine.RenderingMode.OFF_SCREEN;
-
-import com.teamdev.jxbrowser.browser.Browser;
-import com.teamdev.jxbrowser.engine.Engine;
-
-/**
- * This example demonstrates how to load a web page, wait until it is loaded
- * completely, and print its HTML without displaying any GUI.
- */
-public final class HelloWorld {
-
-    public static void main(String[] args) {
-        // Initialize Chromium.
-        Engine engine = Engine.newInstance(OFF_SCREEN);
-
-        // Create a Browser instance.
-        Browser browser = engine.newBrowser();
-
-        // Load a web page and wait until it is loaded completely.
-        browser.navigation().loadUrlAndWait("https://html5test.com/");
-
-        // Print HTML of the loaded web page.
-        browser.mainFrame().ifPresent(frame -> System.out.println(frame.html()));
-
-        // Shutdown Chromium and release allocated resources.
-        engine.close();
-    }
-}
