+++ Gradle/OffScreen/settings.gradle.kts
@@ -18,4 +18,4 @@
  *  OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-rootProject.name = "Core"
+rootProject.name = "OffScreen"
