+++ Maven/OffScreen/src/main/java/HelloOffScreen.java
@@ -18,52 +18,31 @@
  *  OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-import static com.teamdev.jxbrowser.engine.RenderingMode.HARDWARE_ACCELERATED;
+import static com.teamdev.jxbrowser.engine.RenderingMode.OFF_SCREEN;
 
 import com.teamdev.jxbrowser.browser.Browser;
 import com.teamdev.jxbrowser.engine.Engine;
-import com.teamdev.jxbrowser.view.swt.BrowserView;
-import org.eclipse.swt.layout.FillLayout;
-import org.eclipse.swt.widgets.Display;
-import org.eclipse.swt.widgets.Shell;
 
 /**
- * This example demonstrates how to initialize Chromium, create a browser instance
- * (equivalent of the Chromium tab), embed an SWT BrowserView widget into SWT
- * shell to display content of the loaded web page, load the required web page.
+ * This example demonstrates how to load a web page, wait until it is loaded
+ * completely, and print its HTML without displaying any GUI.
  */
-public final class HelloWorld {
+public final class HelloOffScreen {
 
     public static void main(String[] args) {
         // Initialize Chromium.
-        Engine engine = Engine.newInstance(HARDWARE_ACCELERATED);
+        Engine engine = Engine.newInstance(OFF_SCREEN);
 
         // Create a Browser instance.
         Browser browser = engine.newBrowser();
 
-        // Load the required web page.
-        browser.navigation().loadUrl("https://html5test.com");
+        // Load a web page and wait until it is loaded completely.
+        browser.navigation().loadUrlAndWait("https://html5test.com/");
 
-        Display display = new Display();
-        Shell shell = new Shell(display);
-        shell.setText("JxBrowser SWT");
-        shell.setLayout(new FillLayout());
+        // Print HTML of the loaded web page.
+        browser.mainFrame().ifPresent(frame -> System.out.println(frame.html()));
 
-        // Create and embed SWT BrowserView widget to display web content.
-        BrowserView view = BrowserView.newInstance(shell, browser);
-        view.setSize(1280, 800);
-
-        shell.pack();
-        shell.open();
-
-        while (!shell.isDisposed()) {
-            if (!display.readAndDispatch()) {
-                display.sleep();
-            }
-        }
         // Shutdown Chromium and release allocated resources.
         engine.close();
-
-        display.dispose();
     }
 }
